Working on Assembler
Assembler can now translate string expressions into machine codes

+++ Simulator/RegisterHandler.cpp
@@ -9,7 +9,7 @@
 
 RegisterHandler::RegisterHandler() {
     this->registers = (uint32_t *)malloc(sizeof(uint32_t) * 32);
-    this->pc = uint32_t(0);
+    this->pc = 0;
     for(uint8_t i = 0 ; i < 32 ; i++) this->registers[i] = 0;
 }
